refactor: move oauth2 to its own controller
