+++ src/routes/discord/verify.ts
@@ -20,7 +20,7 @@ export class DiscordOAuth2VerifyRoute implements RouteHandler {
 	public async get(req: Request, res: Response) {
 		const { code, state } = req.query as QueryParams;
 		if (!code || !state) return res.status(400).json({ message: 'Missing query parameters' });
-		await this.api.controllers.discord.processOAuth2(code, state);
+		await this.api.controllers.discord.oauth2.processOAuth2(code, state);
 		res.json({ message: 'ok' });
 	}
 }
